Safari-609.1.11.1/LayoutTests/media/video-resize-expected.txt


Test <video> element initial resize event.

EXPECTED (video.videoWidth == '0') OK
EXPECTED (video.videoHeight == '0') OK
EVENT(resize)
EXPECTED (video.videoWidth == '320') OK
EXPECTED (video.videoHeight == '240') OK
END OF TEST
